move telemetry to separate file
